chore: break components into files
