+++ src/components/pages/data/hhs-facilities/definitions.module.scss
@@ -1,0 +1,22 @@
+.definitions-button {
+  @include remove-button-style();
+  text-decoration: underline;
+  color: $color-slate-700;
+  span[aria-hidden] {
+    display: inline-block;
+    text-decoration: none;
+    @include margin(8, left);
+  }
+}
+
+.definitions {
+  display: none;
+  &.is-open {
+    display: block;
+  }
+}
+
+.definitions-wrapper {
+  @include margin(16, top);
+  @include margin(32, bottom);
+}
