+++ src/components/pages/data/hhs-facilities/map/infobox.module.scss
@@ -1,0 +1,27 @@
+.infobox {
+  display: none;
+  @media (hover: hover) {
+    display: block;
+    position: absolute;
+    top: 0;
+    left: 0;
+    z-index: 100;
+    width: 350px;
+    border: 1px solid $color-slate-700;
+    background: white;
+    @include padding(16);
+    @include type-size(100);
+    h3 {
+      @include type-size(100);
+      font-weight: bold;
+      &:first-child {
+        margin-top: 0;
+      }
+    }
+  }
+  p {
+    &:last-child {
+      margin-bottom: 0;
+    }
+  }
+}
